requirements.txt
fastapi
uvicorn
jinja2
python-multipart
sqlalchemy
openai
python-dotenv
websockets